+++ app/src/main/java/feature/film/presentation/FilmsViewModel.kt
@@ -6,6 +6,9 @@ import androidx.lifecycle.ViewModel
 import feature.film.data.FilmsInfra
 import feature.film.domain.models.Film
 import com.example.starwars.models.ScreenState
+import com.example.starwars.infra.FilmsInfra
+import com.example.starwars.models.Film
+import feature.utils.ScreenState
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.rxkotlin.plusAssign
